🗺️ Atlas: [Extract blob lib.rs into modules]
🕸️ Tangle: The `duke-interpreter` crate had a massive 15k+ lines `lib.rs` file containing everything from basic data structures to the core execution loop and registries, demonstrating "The Blob" and "The Sprawl" architectural smells.

📐 Blueprint: Extracted basic state representations into `context.rs` (`ClassContext`, `MethodEntry`, etc.) and registry components into `native_registry.rs`. Kept `lib.rs` as the facade providing `pub use` to maintain backward compatibility without breaking existing module structures. This enforces better domain boundaries.

🧱 Stability: Reduced coupling, faster compile times (as future edits to context won't require recompiling the core registries), and dramatically better readability.

🔬 Verification: Builds successfully (`cargo check --all-targets --all-features`), strict separation enforced, all `cargo test` pass with 0 regressions.
